Make sure pid does not overflow

+++ src/bin/pg_autoctl/pgsetup.c
@@ -437,7 +437,7 @@ get_pgpid(PostgresSetup *pgSetup, bool pg_is_not_running_is_ok)
 
  	fclose(fp);
 
-	if (pid > 0)
+	if (pid > 0 && pid <= INT_MAX)
 	{
 		if (kill(pid, 0) == 0)
 		{
